services/mcp-servers/weather-server/requirements.txt
langchain
langchain_mcp_adapters
httpx
pydantic
python-dotenv
